api: Move to new v1alpha3 API version
In order to better support NodeSelector, switch back to using map[string]string
for Labels. Otherwise NodeSelector and NodeAffinity are not compatible with
LabelSelector.

As this is a breaking change, a new API version is necessary.
Since the whole update to v0.6.0 is breaking, no conversion webhook will be
created.

Signed-off-by: Heathcliff <heathcliff@heathcliff.eu>