upgraded: Load configuration from file when changed
Instead of using node annotations, switch back to reading the configuration
from file.
This allows to use ConfigMaps together with the DaemonSets to fully manage
the upgraded daemons from the controller.
Ensure the loaded config is up-to-date by watching for file changes and
reloading the config when the file changes.

Signed-off-by: Heathcliff <heathcliff@heathcliff.eu>